+++ server/tests/test_apps/accounts/conftest.py
@@ -1,9 +1,9 @@
 from unittest.mock import MagicMock
 
 import pytest
-from the_history_atlas.apps.accounts.database import Database
+from the_history_atlas.apps.accounts.repository import Repository
 
 
 @pytest.fixture
 def mock_db():
-    return MagicMock(spec=Database)
+    return MagicMock(spec=Repository)
