+++ server/the_history_atlas/apps/app_manager.py
@@ -1,4 +1,4 @@
-from the_history_atlas.apps.accounts.accounts import Accounts
+from the_history_atlas.apps.accounts.accounts_app import AccountsApp
 from the_history_atlas.apps.config import Config
 from the_history_atlas.apps.database import DatabaseApp
 from the_history_atlas.apps.history import HistoryApp
@@ -7,13 +7,13 @@ from the_history_atlas.apps.history import HistoryApp
 class AppManager:
     config_app: Config
     database_app: DatabaseApp
-    accounts_app: Accounts
+    accounts_app: AccountsApp
     readmodel_app: HistoryApp
 
     def __init__(self, config_app: Config):
         self.config_app = config_app
         self.database_app = DatabaseApp(config_app=self.config_app)
-        self.accounts_app = Accounts(
+        self.accounts_app = AccountsApp(
             config=self.config_app, database_client=self.database_app.client()
         )
